+++ noodles-util/src/variant.rs
@@ -1,3 +1,6 @@
 //! Variant format utilities.
 
+#[cfg(feature = "async")]
+pub mod r#async;
+
 pub mod io;