+++ noodles-util/src/variant/async.rs
@@ -1,0 +1,3 @@
+//! Async variant format utilities.
+
+pub mod io;