+++ noodles-util/src/variant/async/io.rs
@@ -1,0 +1,5 @@
+//! Async variant format I/O.
+
+mod reader;
+
+pub use self::reader::Reader;